Added actual path to DialNeedle and reimplemented Speedometer
Changed DialNeedle to include actual path for the needle and to adjust the mAngleLimit to allow for changing the division limit.

Uncommented the Speedometer to allow for the right hand dial to be shown and adjusted.

Moved Control to a QWidget to be shown at the bottom of CarPanel. This will be removed in the future. Also changed the signal valueChanged(qreal) to seperate speedChanged and revChanged signals.